Refactor hparams in tests (#1498)
This PR builds on #1491 and completes the transition to removing hparams from unit-testing.
