+++ tests/hparams/test_model_registry.py
