+++ tests/trainer/test_scale_schedule.py
@@ -8,14 +8,15 @@ import pytest
 import torch
 from torch.optim import Optimizer
 from torch.optim.lr_scheduler import ExponentialLR
+from torch.utils.data import DataLoader
 
+from composer import Trainer
 from composer.core import Callback, State, TimeUnit
 from composer.core.types import PyTorchScheduler
 from composer.loggers.logger import Logger
 from composer.optim import MultiStepScheduler
-from composer.optim.optimizer_hparams_registry import SGDHparams
 from composer.trainer._scale_schedule import scale_pytorch_scheduler
-from composer.trainer.trainer_hparams import TrainerHparams
+from tests.common.datasets import RandomClassificationDataset
 from tests.common.models import SimpleModel
 
 
@@ -104,18 +105,19 @@ class TestScaleScheduleTrainer():
     def test_epochs_scaled(
         self,
         ssr: float,
-        composer_trainer_hparams: TrainerHparams,
     ):
+        model = SimpleModel()
+        optimizers = torch.optim.SGD(model.parameters(), lr=1.0)
+        trainer = Trainer(
+            model=model,
+            train_dataloader=DataLoader(RandomClassificationDataset()),
+            optimizers=optimizers,
+            schedulers=[MultiStepScheduler(milestones=['30ba', '50ba'], gamma=0.1)],
+            scale_schedule_ratio=ssr,
+            callbacks=[CheckScaleSchedule(ssr)],
+            max_duration='10ep',
+        )
 
-        composer_trainer_hparams.optimizers = SGDHparams(lr=1.0)
-        composer_trainer_hparams.max_duration = '10ep'
-        composer_trainer_hparams.schedulers = [MultiStepScheduler(milestones=['30ba', '50ba'], gamma=0.1)]
-
-        composer_trainer_hparams.scale_schedule_ratio = ssr
-        trainer = composer_trainer_hparams.initialize_object()
-
-        trainer = composer_trainer_hparams.initialize_object()
-        trainer.state.callbacks.append(CheckScaleSchedule(ssr))
         trainer.state.train_metrics = {}  # avoid metrics construction
 
         assert trainer.state.max_duration is not None
